dom/l10n/tests/mochitest/dom_localization/test_attr_sanitized.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test DOMLocalization's attr sanitization functionality</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript">
  "use strict";
  async function* generateBundles(resourceIds) {
    const bundle = new FluentBundle("en-US");
    bundle.addResource(new FluentResource(`
key1 = Value for Key 1

key2 = Value for <a>Key 2<a/>.
    `));
    yield bundle;
  }

  SimpleTest.waitForExplicitFinish();
  addLoadEvent(async () => {
    const domLoc = new DOMLocalization(
      [],
      false,
      { generateBundles },
    );

    await domLoc.translateFragment(document.body);

    const elem1 = document.querySelector("#elem1");
    const elem2 = document.querySelector("#elem2");

    ok(elem1.textContent.includes("Value for"));
    ok(!elem1.hasAttribute("title"));

    ok(elem2.textContent.includes("Value for"));
    ok(!elem2.hasAttribute("title"));

    SimpleTest.finish();
  });
  </script>
</head>
<body>
  <p id="elem1" title="Old Translation" data-l10n-id="key1"></p>
  <p id="elem2" title="Old Translation" data-l10n-id="key2"></p>
</body>
</html>
